Get rewrite, IsTuple update to more performant version, refactoring
